requirements.txt
torch~=2.5.1+cu121
fastai~=2.8.4
numpy~=2.2.6
pillow~=12.0.0
pycocotools~=2.0.10